zmodem upload的bug
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@3940 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/src/bbs.c
@@ -1960,14 +1960,16 @@ int add_attach(char* file1, char* file2, char* filename)
     char o[8]={0,0,0,0,0,0,0,0};
     char buf[1024*16];
     int i;
-    stat(file2, &st);
+    if(stat(file2, &st)==-1)
+        return 0;
     if(st.st_size>=1024*1024&&!HAS_PERM(currentuser, PERM_SYSOP)) {
         unlink(file2);
         return 0;
     }
     size=htonl(st.st_size);
-    fp=fopen(file1, "ab");
     fp2=fopen(file2, "rb");
+    if(!fp2) return 0;
+    fp=fopen(file1, "ab");
     fwrite(o,1,8,fp);
     for(i=0;i<strlen(filename);i++)
         if(!isalnum(filename[i])&&filename[i]!='.') filename[i]='A';
